+++ firebase-vscode/src/utils/settings.ts
@@ -10,6 +10,7 @@ export interface Settings {
   readonly importPath?: string;
   readonly exportPath: string;
   readonly exportOnExit: boolean;
+  readonly debug: boolean;
 }
 
 // TODO: Temporary fallback for bashing, this should probably point to the global firebase binary on the system
@@ -43,6 +44,7 @@ export function getSettings(): Settings {
     importPath: config.get<string>("emulators.importPath"),
     exportPath: config.get<string>("emulators.exportPath", "./exportedData"),
     exportOnExit: config.get<boolean>("emulators.exportOnExit", false),
+    debug: config.get<boolean>("debug", false),
   };
 }