+++ firebase-vscode/src/analytics.ts
@@ -32,6 +32,7 @@ export enum DATA_CONNECT_EVENT_NAME {
   START_EMULATORS = "start_emulators",
   AUTO_COMPLETE = "auto_complete",
   SESSION_CHAR_COUNT = "session_char_count",
+  EMULATOR_EXPORT ="emulator_export",
   SETUP_FIREBASE_BINARY = "setup_firebase_binary",
 }