+++ firebase-vscode/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.12.1
 
-- Updated inertnal firebase-tools dependency to 13.29.2
+- Updated internal `firebase-tools` dependency to 13.29.2
 - [Added] Added support for emulator import/export.
 - [Added] Added `debug` setting to run commands with `--debug`
 - [Fixed] Fixed a bug where emulator issues weren't being surfaced